backend/requirements.txt
# =============================================================================
# Django Project Tracking Management System - Python Dependencies
# =============================================================================

# Core Django
Django>=5.0,<5.1

# Django REST Framework
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0

# Authentication & Security
django-bcrypt>=0.9.2
bcrypt>=4.1.0

# API Documentation
drf-spectacular>=0.27.0

# Geospatial
geopy>=2.4.0

# Database
mysqlclient>=2.2.0

# Caching & Sessions
django-redis>=5.4.0
redis>=5.0.0

# Background Tasks
Celery>=5.3.0
celery-redbeat>=2.2.0

# WebSocket
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
Pillow>=10.1.0

# Logging
structlog>=23.2.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
black>=23.12.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.7.0
django-stubs>=4.2.0
djangorestframework-stubs>=3.14.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0
sentry-sdk>=1.39.0
